[HT-20] Added new layout for the list of habits
The layout for the list of cards was added, habit_list.xml. The fragment HabitListFragment was added, in order to substitute Activity's container with fragment layout. The layout habit_card was added as an element of habits' list. The adapter HabitListAdapter for the list of habits was added, with dummy data in it.